Update several files

+++ src/layers/services/room.service.js
@@ -7,20 +7,34 @@ class RoomService {
         this.roomRepository = new RoomRepository();
     }
 
-    postRoom = async (roomName, category, ownerUserId) => {
-        return await this.roomRepository.postRoom(roomName, category, ownerUserId);
+    errResponse = (status, message) => {
+        return { status, success: false, message };
     };
 
-    getRoom = async (req, res, next) => {
+    postRoom = async (ownerUserId, roomName, content, category) => {
+        return await this.roomRepository.postRoom(ownerUserId, roomName, content, category);
+    };
+
+    getRoom = async () => {
         return await this.roomRepository.getRoom();
     };
 
     getCategoryRoom = async (category) => {
-        return await this.roomRepository.getCategoryRoom(category);
+        const getCategoryRoom = await this.roomRepository.getCategoryRoom(category);
+        if (!getCategoryRoom) return this.errResponse(401, "알 수 없는 에러");
+
+        return { status: 200, success: true, result: getCategoryRoom };
     };
 
-    deleteRoom = async (roomId) => {
-        return await this.roomRepository.deleteRoom(roomId);
+    deleteRoom = async (roomId, ownerUserId) => {
+        const roomInfo = await this.roomRepository.getOneRoom(roomId);
+        if (!roomInfo) return this.errResponse(400, "존재하지 않는 룸입니다.");
+        if (roomInfo.ownerUserId !== ownerUserId) return this.errResponse(401, "방장이 아닙니다.");
+
+        const deleteRoom = await this.roomRepository.deleteRoom(roomId);
+        if (!deleteRoom) return this.errResponse(401, "알 수 없는 에러");
+
+        return { status: 200, success: true, result: {} };
     };
 }
 module.exports = RoomService;
